+++ src/f3/go_ffi.rs
@@ -13,6 +13,7 @@ pub trait GoF3Node {
         rpc_endpoint: String,
         f3_rpc_endpoint: String,
         initial_power_table: String,
+        bootstrap_epoch: i64,
         finality: i64,
         f3_root: String,
         manifest_server: String,